feat: integration tests

+++ Backend/src/controllers/EscolaController.js
@@ -4,9 +4,9 @@ export const createEscola = async (req, res) => {
     try {
         const newData = req.body;
         const escola = await EscolaService.createEscola(newData);
-        return res.status(200).json(escola);
+        return res.status(201).json(escola);
     }catch (error){
-        return res.status(404).json({message: error.message});
+        return res.status(404).json({error: error.message});
     }
 };
 
@@ -16,7 +16,7 @@ export const getAllEscolas = async (req, res) => {
         const escolas = await EscolaService.getAllEscolas(search);
         return res.status(200).json(escolas);
     } catch (error) {
-        return res.status(404).json({message: error.message});
+        return res.status(404).json({error: error.message});
     }
 };
 
@@ -26,7 +26,7 @@ export const getAllTurmasByEscolaId = async (req, res) => {
         const turmas = await EscolaService.getAllTurmasByEscolaId(id);
         return res.status(200).json(turmas);
     } catch (error) {
-        return res.status(404).json({message: error.message});
+        return res.status(404).json({error: error.message});
     }
 };
 
@@ -36,7 +36,7 @@ export const getAllTurmasByEscolaId = async (req, res) => {
 //         const escolas = await EscolaService.getAllEscolas(search);
 //         return res.status(200).json(escolas);
 //     } catch (error) {
-//         return res.status(404).json({message: error.message});
+//         return res.status(404).json({error: error.message});
 //     }
 // };
 
@@ -44,9 +44,12 @@ export const getEscolaById = async (req, res) => {
     try {
         const {id} = req.params;
         const escola = await EscolaService.getEscolaById(id);
+        if(!escola){
+            return res.status(404).json({ error: "Escola não encontrada" });
+        }
         return res.status(200).json(escola);
     } catch (error) {
-        return res.status(404).json({message: error.message});
+        return res.status(404).json({error: error.message});
     }
 };
 
@@ -54,20 +57,28 @@ export const updateEscola = async (req, res) => {
     try {
         const {id} = req.params;
         const data = req.body;
+        const escola = await EscolaService.getEscolaById(id);
+        if (!escola) {
+            return res.status(404).json({ error: "Escola não encontrada" });
+        }
         const updatedEscola = await EscolaService.updateEscola(id, data);
         return res.status(200).json(updatedEscola);
     } catch (error) {
-        return res.status(404).json({ message: error.message });
+        return res.status(404).json({error: error.message});
     }
 };
 
 export const deleteEscola = async (req, res) => {
     try {
         const {id} = req.params;
-        await EscolaService.deleteEscola(id);
-        return res.status(200).json({message: 'Escola apagada' });
+        const escola = await EscolaService.getEscolaById(id);
+        if (!escola) {
+            return res.status(404).json({ error: "Escola não encontrada" });
+        }
+        const result = await EscolaService.deleteEscola(id);
+        return res.status(200).json(result); 
     } catch (error) {
-        return res.status(404).json({message: error.message});
+        return res.status(404).json({error: error.message});
     }
 };
 
@@ -82,7 +93,7 @@ export const createCartaConvite = async (req, res) => {
         });
         return res.send(pdfBuffer);
     } catch (error) {
-        return res.status(404).json({ message: error.message });
+        return res.status(404).json({error: error.message});
     }
 };
 
@@ -97,6 +108,6 @@ export const createCartaConvenio = async (req, res) => {
         });
         return res.send(pdfBuffer);
     } catch (error) {
-        return res.status(404).json({ message: error.message });
+        return res.status(404).json({error: error.message});
     }
 };
